+++ mediation/DTExchange/source/plugin/Assets/GoogleMobileAds/Mediation/DTExchange/Platforms/Mediation/DTExchangeClientFactory.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2019 Google LLC
+// Copyright 2019 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -21,13 +21,13 @@ namespace GoogleMobileAds.Mediation.DTExchange
         public static IDTExchangeClient CreateDTExchangeClient()
         {
             #if UNITY_EDITOR
-            return new GoogleMobileAds.Mediation.DTExchange.Common.DummyClient();
+            return new GoogleMobileAds.Mediation.DTExchange.Common.PlaceholderClient();
             #elif UNITY_ANDROID
             return GoogleMobileAds.Mediation.DTExchange.Android.DTExchangeClient.Instance;
             #elif UNITY_IOS
             return GoogleMobileAds.Mediation.DTExchange.iOS.DTExchangeClient.Instance;
             #else
-            return new GoogleMobileAds.Mediation.DTExchange.Common.DummyClient();
+            return new GoogleMobileAds.Mediation.DTExchange.Common.PlaceholderClient();
             #endif
         }
     }